Refactor ReactSelectStyles, moving specific functions to the components using them.
